Extract risk rolling contracts
